move parser from main server file to utils; fixed some parser bugs; added parser tests
